refactor: /simplify cleanup pass on the health-refactor diff
Quality cleanups found by a 4-angle review of this PR's own diff
(reuse / simplification / efficiency / altitude). Behavior unchanged.

- engine-install.ts: validateBackend/validateDiarize now take caps as a
  param; downloadEngine probes getEngineCapabilities() once behind a single
  shared gate instead of each helper fetching independently.
- install-plan.ts: buildTtsComponents returns wantsAnyKokoro so renderInstallPlan
  reads tts.wantsAnyKokoro instead of re-deriving the same predicate.
- synth.ts: inline single-use supportsAcronymExpansion predicate into its only
  caller (kept as a named local) — removes a one-caller indirection layer.
- status.ts: logEngineCapabilities uses getEngineCapabilities().catch(() => null)
  instead of let-null + try/catch resetting to the same sentinel.
- cli/types.ts: moved SharedInstallArgs out of install.ts into a shared types
  module so init.ts no longer imports a type from its sibling command.

Skipped: 3x humanBytes duplication (pre-existing, outside this diff, impls
intentionally diverge) and the status disk-helper return-shape restructure
(judgment call, output-drift risk).
