+++ src/cli/install.ts
@@ -7,23 +7,7 @@ import { maybeAskForStar } from "../star";
 import { log } from "../log";
 import { packageVersion } from "../package-info";
 import { createDiagnosticLogSession, type DiagnosticLogSession } from "../diagnostic-log";
-
-/**
- * Fields shared between `install` and `init` command args. Both commands
- * accept the same backend/cache/model flags; extracting them here avoids
- * duplicating the declarations in `init.ts`.
- */
-export interface SharedInstallArgs {
-  coreml: boolean;
-  onnx: boolean;
-  "no-cache": boolean;
-  noCache?: boolean;
-  no_cache?: boolean;
-  tts: boolean;
-  vad: boolean;
-  diarize: boolean;
-  plan: boolean;
-}
+import type { SharedInstallArgs } from "./types";
 
 export interface InstallCommandArgs extends SharedInstallArgs {
   /** Positional args after `install` — candidate TTS language codes. */
